+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Click a line to open the map with that mob focused
     * Control-click a line to set a waypoint for the closest location for that mob
     * Shift-click a line to link the closest location for that mob in chat
+    * Mousing over them triggers the highlights on the map, if it's open
 * Minimap pins update their focus state immediately
 
 ## Changed in v90001.5