+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Shift-click a line to link the closest location for that mob in chat
     * Mousing over them triggers the highlights on the map, if it's open
 * Minimap pins update their focus state immediately
+* Fix a spawn location for Krik'thir the Gatewatcher
 
 ## Changed in v90001.5